# expression: refactor cache logic for `builtinRegexpSubstrFuncSig`