--- a/pkg/expression/builtin_regexp_test.go
+++ b/pkg/expression/builtin_regexp_test.go
@@ -1224,3 +1224,80 @@ func TestRegexpReplaceVec(t *testing.T) {
 
 	testVectorizedBuiltinFunc(t, vecBuiltinRegexpReplaceCases)
 }
+
+func TestRegexpCache(t *testing.T) {
+	ctx := createContext(t)
+
+	// if the pattern or match type is not constant, it should not be cached
+	sig := regexpNewBaseFuncSig{}
+	sig.args = []Expression{&Column{}, &Column{}, &Constant{}}
+	reg, err := sig.getRegexp(ctx, "abc", "", 2)
+	require.NoError(t, err)
+	require.Equal(t, "abc", reg.String())
+
+	reg, err = sig.getRegexp(ctx, "def", "", 2)
+	require.NoError(t, err)
+	require.Equal(t, "def", reg.String())
+
+	reg, ok, err := sig.tryVecMemorizedRegexp(ctx, []*funcParam{
+		{defaultStrVal: "x"},
+		{defaultStrVal: "aaa"},
+		{defaultStrVal: ""},
+	}, 2, 1)
+	require.Nil(t, reg)
+	require.False(t, ok)
+	require.NoError(t, err)
+
+	_, ok = sig.memorizedRegexp.getCache(ctx.GetSessionVars().StmtCtx.CtxID())
+	require.False(t, ok)
+
+	sig.args = []Expression{&Column{}, &Constant{}, &Column{}}
+	reg, err = sig.getRegexp(ctx, "bbb", "", 2)
+	require.NoError(t, err)
+	require.Equal(t, "bbb", reg.String())
+
+	reg, ok, err = sig.tryVecMemorizedRegexp(ctx, []*funcParam{
+		{defaultStrVal: "x"},
+		{defaultStrVal: "aaa"},
+		{defaultStrVal: ""},
+	}, 2, 1)
+	require.Nil(t, reg)
+	require.False(t, ok)
+	require.NoError(t, err)
+
+	_, ok = sig.memorizedRegexp.getCache(ctx.GetSessionVars().StmtCtx.CtxID())
+	require.False(t, ok)
+
+	// if pattern and match type are both constant, it should be cached
+	sig = regexpNewBaseFuncSig{}
+	sig.args = []Expression{&Column{}, &Constant{ParamMarker: &ParamMarker{}}, &Constant{ParamMarker: &ParamMarker{}}}
+	reg, err = sig.getRegexp(ctx, "ccc", "", 2)
+	require.NoError(t, err)
+	require.Equal(t, "ccc", reg.String())
+
+	reg2, err := sig.getRegexp(ctx, "ddd", "", 2)
+	require.NoError(t, err)
+	require.Same(t, reg, reg2)
+	require.Equal(t, "ccc", reg2.String())
+
+	sig = regexpNewBaseFuncSig{}
+	sig.args = []Expression{&Column{}, &Constant{ParamMarker: &ParamMarker{}}, &Constant{ParamMarker: &ParamMarker{}}}
+	reg, ok, err = sig.tryVecMemorizedRegexp(ctx, []*funcParam{
+		{defaultStrVal: "x"},
+		{defaultStrVal: "ddd"},
+		{defaultStrVal: ""},
+	}, 2, 1)
+	require.Equal(t, "ddd", reg.String())
+	require.True(t, ok)
+	require.NoError(t, err)
+
+	reg2, ok, err = sig.tryVecMemorizedRegexp(ctx, []*funcParam{
+		{defaultStrVal: "x"},
+		{defaultStrVal: "eee"},
+		{defaultStrVal: ""},
+	}, 2, 1)
+	require.Same(t, reg, reg2)
+	require.Equal(t, "ddd", reg2.String())
+	require.True(t, ok)
+	require.NoError(t, err)
+}