--- a/tests/integrationtest/r/expression/plan_cache.result
+++ b/tests/integrationtest/r/expression/plan_cache.result
@@ -94,16 +94,49 @@ set tidb_enable_prepared_plan_cache=ON;
 drop table if exists t1;
 create table t1 (a varchar(40));
 insert into t1 values ('C1'),('R1');
+insert into mysql.expr_pushdown_blacklist values('regexp_like', 'tikv,tiflash,tidb', 'for test');
+insert into mysql.expr_pushdown_blacklist values('regexp_substr', 'tikv,tiflash,tidb', 'for test');
+admin reload expr_pushdown_blacklist;
+set tidb_enable_vectorized_expression=ON;
 prepare stmt1 from 'select a from t1 where a rlike ?';
+prepare stmt2 from 'select a, regexp_substr(a, ?) from t1 order by a asc';
 set @a='^C.*';
 execute stmt1 using @a;
 a
 C1
+execute stmt2 using @a;
+a	regexp_substr(a, ?)
+C1	C1
+R1	NULL
 set @a='^R.*';
 execute stmt1 using @a;
 a
 R1
+execute stmt2 using @a;
+a	regexp_substr(a, ?)
+C1	NULL
+R1	R1
+set tidb_enable_vectorized_expression=OFF;
+set @a='^C.*';
+execute stmt1 using @a;
+a
+C1
+execute stmt2 using @a;
+a	regexp_substr(a, ?)
+C1	C1
+R1	NULL
+set @a='^R.*';
+execute stmt1 using @a;
+a
+R1
+execute stmt2 using @a;
+a	regexp_substr(a, ?)
+C1	NULL
+R1	R1
+delete from mysql.expr_pushdown_blacklist where name like 'regexp%' and store_type = 'tikv,tiflash,tidb' and reason = 'for test';
+admin reload expr_pushdown_blacklist;
 set tidb_enable_prepared_plan_cache=default;
+set tidb_enable_vectorized_expression=default;
 set tidb_enable_prepared_plan_cache=ON;
 drop table if exists t;
 create table t(col_int int);