<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #49583

Problem Summary:

See #49583 and #49464

### What changed and how does it work?

1. Add `CtxID` to `StatementContext` to return a unique ID for statement context. 
2. Introduce `builtinFuncCache` to provide a cache mechanism for built function. It uses context ID to separate cache between different statements so that we can still cache some immediate states even if plan cache is enabled now.
3. Introduce `regexpNewBaseFuncSig` as a new base for regex related functions. We'll use `regexpNewBaseFuncSig` to replace the old `regexpBaseFuncSig` in the next several PRs.
4. refactor `builtinRegexpSubstrFuncSig` to use `regexpNewBaseFuncSig` to manage the cache, so that:
  - It's implement can decouple with plan cache
  - It can still cache the parsed `*regexp.Regexp` when plan cache is on and regex pattern is read from a variable

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [x] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->


We did a benchmark test and the performance of `regexp_substr` increased:

```
func BenchmarkRegexpSubStr(b *testing.B) {
	store := testkit.CreateMockStore(b)
	tk := testkit.NewTestKit(b, store)
	tk.MustExec("use test")
	tk.MustExec("create table t(a varchar(256))")
	for i := 0; i < b.N; i++ {
		tk.MustExec(fmt.Sprintf("insert into t values('pingcaper%d@pingcap.com')", i))
	}
	tk.MustExec(`set @regexp='^\\w+([-+.]\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*$'`)
	tk.MustExec("prepare s from 'select regexp_substr(a, ?) from t'")

	b.ResetTimer()
	tk.MustQuery("execute s using @regexp")
	b.StopTimer()
}
```

Master:

```
go test -bench='^BenchmarkRegexpSubStr$' -run='^$' -benchtime=1000000x --tags=intest
...
 1000000	      1804 ns/op
PASS
ok  	github.com/pingcap/tidb/pkg/expression/integration_test	64.611s
```

This PR:
```
go test -bench='^BenchmarkRegexpSubStr$' -run='^$' -benchtime=1000000x --tags=intest
...
 1000000	       333.3 ns/op
PASS
ok  	github.com/pingcap/tidb/pkg/expression/integration_test	64.772s
```

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
improve the performance of function `regexp_substr` in some cases when plan cache enabled
```